Add files via upload
This project was a virtual experience program in the field of Data Science. 

The first achievement contained the Exploratory data analysis wherein data cleaning and transformations were applied on the dataset provided by the bank ANZ (containing the details of the transactions within a 3 month period). A proper analysis was created which was visualised in 2 parts using Tableau, the first being the transactions which were posted and the second corresponding to the authorised transactions. Two visually appealing dashboards were created which can be viewed using the files with extensions '.hyper' . 

The second achievement depicts the Predictive data analysis, wherein I ran 2 separate models (Linear Regression and Decision Tree Regression) on the provided dataset and predicted the annual salaries for 100 customers. Linear regression was completed by achieving a Mean Square error of 0.034 and the Decision Tree Regression model achieved 0.039 value for the same.